+++ game/room/hidden_area_manager.gd
@@ -1,0 +1,40 @@
+# tabletop-club
+# Copyright (c) 2020-2023 Benjamin 'drwhut' Beddows.
+# Copyright (c) 2021-2023 Tabletop Club contributors (see game/CREDITS.tres).
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
+
+class_name HiddenAreaManager
+extends IndexManager
+
+## Manages the hidden areas that players can add and remove during gameplay.
+##
+## TODO: Test this class.
+
+
+# The PackedScene containing the hidden area.
+var _hidden_area_scene := preload("res://room/hidden_area.tscn")
+
+
+## Add a hidden area with the given index and transform to the room.
+## TODO: Add player information.
+func add_hidden_area(index: int, transform: Transform) -> void:
+	var hidden_area = _hidden_area_scene.instance()
+	hidden_area.transform = transform
+	add_child_with_index(index, hidden_area)