+++ CHANGELOG.md
@@ -89,6 +89,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   loading screen when entering singleplayer or multiplayer.
 - The in-game menu screen now uses the same layout as the main menu.
 - Optimised the performance of the paint and erase tools.
+- The appearance of hidden areas is now animated.
 
 ### Fixed